codemod: split jest-to-vitest into a phased pipeline, export each step, add shared TSX parser helpers, flatten plop codemod output path
