+++ packages/codemod/package.json
@@ -38,6 +38,15 @@
   "types": "./dist/index.d.mts",
   "exports": {
     "./codemods/jest-to-vitest": "./dist/codemods/jest-to-vitest.mjs",
+    "./codemods/jest-to-vitest-failing": "./dist/codemods/jest-to-vitest-failing.mjs",
+    "./codemods/jest-to-vitest-fn-generics": "./dist/codemods/jest-to-vitest-fn-generics.mjs",
+    "./codemods/jest-to-vitest-hooks": "./dist/codemods/jest-to-vitest-hooks.mjs",
+    "./codemods/jest-to-vitest-imports": "./dist/codemods/jest-to-vitest-imports.mjs",
+    "./codemods/jest-to-vitest-jest-members": "./dist/codemods/jest-to-vitest-jest-members.mjs",
+    "./codemods/jest-to-vitest-mock-factories": "./dist/codemods/jest-to-vitest-mock-factories.mjs",
+    "./codemods/jest-to-vitest-mock-types": "./dist/codemods/jest-to-vitest-mock-types.mjs",
+    "./codemods/jest-to-vitest-resolves-rejects-generics": "./dist/codemods/jest-to-vitest-resolves-rejects-generics.mjs",
+    "./codemods/jest-to-vitest-set-timeout": "./dist/codemods/jest-to-vitest-set-timeout.mjs",
     "./codemods/svg-import-query-param": "./dist/codemods/svg-import-query-param.mjs",
     "./codemods/transform-vite-loadable": "./dist/codemods/transform-vite-loadable.mjs",
     "./program": "./dist/program/index.mjs",
